Extract WooPosBookingInfoLoader and remove WooPosGetLineItemBookingIds
Move booking cache check, network fetch, and formatting into
WooPosBookingInfoLoader. Simplify WooPosOrderDetailsMapper to
use item.bookingId directly instead of WCOrderStore lookup.
Simplify ViewModel sideLoadBookings by delegating to the loader.
