+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/orders/details/WooPosBookingInfoLoader.kt
@@ -1,0 +1,55 @@
+package com.woocommerce.android.ui.woopos.orders.details
+
+import com.woocommerce.android.R
+import com.woocommerce.android.ui.bookings.Booking
+import com.woocommerce.android.ui.bookings.BookingsRepository
+import com.woocommerce.android.ui.woopos.orders.WooPosOrdersState.OrderDetailsViewState.Computed.Details.BookingInfo
+import com.woocommerce.android.util.DateFormatter
+import com.woocommerce.android.viewmodel.ResourceProvider
+import java.time.ZoneOffset
+import java.time.format.DateTimeFormatter
+import java.time.format.FormatStyle
+import javax.inject.Inject
+
+class WooPosBookingInfoLoader @Inject constructor(
+    private val bookingsRepository: BookingsRepository,
+    private val resourceProvider: ResourceProvider,
+    private val dateFormatter: DateFormatter,
+) {
+    suspend fun resolveBookingInfo(bookingId: Long): BookingInfo {
+        val booking = bookingsRepository.getBooking(bookingId)
+        return if (booking != null) {
+            BookingInfo.Loaded(formatBookingInfo(bookingId, booking))
+        } else {
+            BookingInfo.Loading(bookingId)
+        }
+    }
+
+    suspend fun fetchBookingInfo(bookingId: Long): BookingInfo {
+        return bookingsRepository.fetchBooking(bookingId).fold(
+            onSuccess = { booking ->
+                BookingInfo.Loaded(formatBookingInfo(bookingId, booking))
+            },
+            onFailure = {
+                BookingInfo.Error(
+                    resourceProvider.getString(R.string.woopos_orders_details_booking_info_error)
+                )
+            }
+        )
+    }
+
+    private fun formatBookingInfo(bookingId: Long, booking: Booking): String {
+        val dateFormat = DateTimeFormatter.ofLocalizedDate(FormatStyle.MEDIUM)
+            .withZone(ZoneOffset.UTC)
+        val date = dateFormat.format(booking.start)
+        val startTime = dateFormatter.formatTime(booking.start)
+        val endTime = dateFormatter.formatTime(booking.end)
+        return resourceProvider.getString(
+            R.string.woopos_orders_details_booking_info,
+            bookingId,
+            date,
+            startTime,
+            endTime
+        )
+    }
+}
